requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.9
diff-match-patch==20241021
distro==1.9.0
Django==5.2.4
django-debug-toolbar==5.2.0
django-extensions==4.1
django-grappelli==4.0.2
django-import-export==4.3.8
django-jazzmin==3.0.1
django-simple-history==3.10.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.95.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
